Move `MobilizonEvent.from_model` to `storage.query`
