+++ mobilizon_reshare/cli/commands/format/format.py
@@ -3,6 +3,7 @@ import click
 from mobilizon_reshare.event.event import MobilizonEvent
 from mobilizon_reshare.models.event import Event
 from mobilizon_reshare.publishers.platforms.platform_mapping import get_formatter_class
+from mobilizon_reshare.storage.query import from_model
 
 
 async def format_event(event_id, publisher_name: str):
@@ -12,6 +13,6 @@ async def format_event(event_id, publisher_name: str):
     if not event:
         click.echo(f"Event with mobilizon_id {event_id} not found.")
         return
-    event = MobilizonEvent.from_model(event)
+    event = from_model(event)
     message = get_formatter_class(publisher_name)().get_message_from_event(event)
     click.echo(message)
