+++ tests/commands/test_start.py
@@ -4,10 +4,10 @@ from logging import DEBUG, INFO
 import arrow
 import pytest
 
-from mobilizon_reshare.storage.query import to_model
+from mobilizon_reshare.storage.query import to_model, from_model
 from mobilizon_reshare.storage.query.read import get_all_events
 from tests.commands.conftest import simple_event_element
-from mobilizon_reshare.event.event import MobilizonEvent, EventPublicationStatus
+from mobilizon_reshare.event.event import EventPublicationStatus
 from mobilizon_reshare.main.start import start
 from mobilizon_reshare.models.event import Event
 from mobilizon_reshare.models.publication import PublicationStatus
@@ -75,7 +75,7 @@ async def test_start_new_event(
 
         # the derived status for the event should be COMPLETED
         assert (
-            MobilizonEvent.from_model(all_events[0]).status
+            from_model(all_events[0]).status
             == EventPublicationStatus.COMPLETED
         )
 
@@ -119,10 +119,7 @@ async def test_start_event_from_db(
             assert p.status == PublicationStatus.COMPLETED
 
         # the derived status for the event should be COMPLETED
-        assert (
-            MobilizonEvent.from_model(event_model).status
-            == EventPublicationStatus.COMPLETED
-        )
+        assert from_model(event_model).status == EventPublicationStatus.COMPLETED
 
 
 @pytest.mark.asyncio
@@ -169,10 +166,7 @@ async def test_start_publisher_failure(
             for _ in range(2)
         ]  # 2 publications failed * 2 notifiers
         # the derived status for the event should be FAILED
-        assert (
-            MobilizonEvent.from_model(event_model).status
-            == EventPublicationStatus.FAILED
-        )
+        assert from_model(event_model).status == EventPublicationStatus.FAILED
 
 
 @pytest.fixture
